+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 #### Added
 
+- Split the Project index page into two tabs - "In Progress" and "Completed"
+  projects
 - Set up selenium to work within CI
 - add in accessibility page
 - footer now have link to accessibility page